USING GOOGLE SHEET
Automate the scaling process without a scale operator. It is a system that will guarantees 99 to 100% accurate count and you are in complete control.

The person who inspects the raw material in the container will be there and the forklift driver will be needed as well.

All the info will be contained in the barcode and all the info contained in the barcode will determine if you get an accurate count. The barcode will contain the weight, the type of container, the type of raw material in the container, and more.

Google sheet should be used because anyone can read the data as long as they have access to the internet. You get an instant result and you can use MACROS to automate the input, calculation, and output process.